Fix setup (#54)
When running the setup, the symbolic link is failing due to the
target destination folder not existing on a clean environment.

+++ setup.sh
@@ -34,11 +34,10 @@ function setup_defaults
     fi
     WM_UTS_JLINK_INSTALL_PATH=${WM_UTS_JLINK_INSTALL_PATH:-"${HOME}/opt"}
 
-
     #todo: write it to user.env
-    mkdir -p ${WM_UTS_SERVICE_HOME}
-    mkdir -p ${WM_UTS_JLINK_INSTALL_PATH}
-    mkdir -p ${WM_UTS_GCC_INSTALL_PATH}
+    mkdir -p "${WM_UTS_SERVICE_HOME}"
+    mkdir -p "${WM_UTS_JLINK_INSTALL_PATH}"
+    mkdir -p "${WM_UTS_GCC_INSTALL_PATH}"
 
     export PATH=${PATH}:${WM_UTS_INSTALL_PATH}
 }
@@ -52,8 +51,8 @@ function update_path_config
     _VALUE=${2}
 
     echo "removing ${_KEY} from env files"
-    sed -i "/${_KEY}/d" ${WM_UTS_ENVIRONMENT_CUSTOM}
-    echo "${_KEY}=${_VALUE}" >> ${WM_UTS_ENVIRONMENT_CUSTOM}
+    sed -i "/${_KEY}/d" "${WM_UTS_ENVIRONMENT_CUSTOM}"
+    echo "${_KEY}=${_VALUE}" >> "${WM_UTS_ENVIRONMENT_CUSTOM}"
 
 }
 
@@ -64,9 +63,9 @@ function setup_gcc
 
     if [[ ! -f ${GCC_ARCHIVE} ]]
     then
-        mkdir -p ${WM_UTS_GCC_INSTALL_PATH}
-        wget ${_gcc_url}
-        tar -xjvf ${GCC_ARCHIVE} -C ${WM_UTS_GCC_INSTALL_PATH}
+        mkdir -p "${WM_UTS_GCC_INSTALL_PATH}"
+        wget "${_gcc_url}"
+        tar -xjvf "${GCC_ARCHIVE}" -C "${WM_UTS_GCC_INSTALL_PATH}"
     fi
 
     export PATH="${WM_UTS_GCC_INSTALL_PATH}/${WM_UTS_GCC_VERSION}/bin:$PATH"
@@ -74,21 +73,21 @@ function setup_gcc
 
 function setup_jlink
 {
-    if [[ ! -f ${JLINK_ARCHIVE} ]]
+    if [[ ! -f "${JLINK_ARCHIVE}" ]]
     then
-        mkdir -p ${WM_UTS_JLINK_INSTALL_PATH}
+        mkdir -p "${WM_UTS_JLINK_INSTALL_PATH}"
         echo "Please download SEGGER jlink:"
         echo "https://www.segger.com/downloads/jlink/${JLINK_ARCHIVE}"
         echo ""
         read -n 1 -s -r -p "Press any key to continue"
         echo ""
-        if [[ ! -f ${JLINK_ARCHIVE} ]]
+        if [[ ! -f "${JLINK_ARCHIVE}" ]]
         then
             echo "Please move the archive to $(pwd)"
             read -n 1 -s -r -p "Press any key to continue"
             echo ""
         fi
-        tar -xvzf ${JLINK_ARCHIVE} -C ${WM_UTS_JLINK_INSTALL_PATH}
+        tar -xvzf "${JLINK_ARCHIVE}" -C "${WM_UTS_JLINK_INSTALL_PATH}"
     fi
 
     export PATH="${WM_UTS_JLINK_INSTALL_PATH}:$PATH"
@@ -114,17 +113,17 @@ function setup_python
 
 function setup_wmutils
 {
-    if [[ -f ${WM_UTS_SERVICE_HOME}/bin/wm-utils.sh  ]]
+    if [[ -f "${WM_UTS_SERVICE_HOME}/bin/wm-utils.sh"  ]]
     then
         echo "copying exec from bundle"
         _is_clone=false
-    elif [[ -f $(pwd)/bin/wm-utils.sh  ]]
+    elif [[ -f "$(pwd)/bin/wm-utils.sh"  ]]
     then
         echo "assuming git clone: setting up ${WM_UTS_SERVICE_HOME}"
         export BUILD_VERSION=$(git log -n 1 --oneline --format=%h)
 
         echo "setting up ${WM_UTS_SERVICE_HOME}"
-        rsync -av  . ${WM_UTS_SERVICE_HOME} \
+        rsync -av  . "${WM_UTS_SERVICE_HOME}" \
               --exclude .git \
               --exclude setup.sh \
               --exclude *tar* \
@@ -135,28 +134,25 @@ function setup_wmutils
         _is_clone=true
     fi
 
-    update_path_config WM_UTS_GCC_INSTALL_PATH ${WM_UTS_GCC_INSTALL_PATH}/${WM_UTS_GCC_VERSION}/bin
-    update_path_config WM_UTS_JLINK_INSTALL_PATH ${WM_UTS_JLINK_INSTALL_PATH}/${WM_UTS_JLINK_VERSION}
+    update_path_config WM_UTS_GCC_INSTALL_PATH "${WM_UTS_GCC_INSTALL_PATH}/${WM_UTS_GCC_VERSION}/bin"
+    update_path_config WM_UTS_JLINK_INSTALL_PATH "${WM_UTS_JLINK_INSTALL_PATH}/${WM_UTS_JLINK_VERSION}"
 
     # copy and set permissions
-    if [[ ${INSTALL_SYSTEM_WIDE} == true ]]
+    if [[ "${INSTALL_SYSTEM_WIDE}" == true ]]
     then
         echo "installing wm-utils system wide"
-        sudo ln -s ${WM_UTS_SERVICE_HOME}/bin/wm-utils.sh ${WM_UTS_INSTALL_PATH}/${WM_UTS_EXEC_NAME}
-        sudo chmod +x ${WM_UTS_INSTALL_PATH}/${WM_UTS_EXEC_NAME}
+        sudo mkdir -vp "${WM_UTS_INSTALL_PATH}"
+        sudo ln -vfs "${WM_UTS_SERVICE_HOME}/bin/wm-utils.sh" "${WM_UTS_INSTALL_PATH}/${WM_UTS_EXEC_NAME}"
+        sudo chmod +x "${WM_UTS_INSTALL_PATH}/${WM_UTS_EXEC_NAME}"
     else
         echo "installing wm-utils for current user"
-
-        rm -Rf ${WM_UTS_INSTALL_PATH}/${WM_UTS_EXEC_NAME}
-
-        mkdir -p /home/${USER}/.local/bin || true
-        ln -s ${WM_UTS_SERVICE_HOME}/bin/wm-utils.sh ${WM_UTS_INSTALL_PATH}/${WM_UTS_EXEC_NAME}
-        chmod +x ${WM_UTS_INSTALL_PATH}/${WM_UTS_EXEC_NAME}
-
+        mkdir -vp "${WM_UTS_INSTALL_PATH}"
+        ln -vfs "${WM_UTS_SERVICE_HOME}/bin/wm-utils.sh" "${WM_UTS_INSTALL_PATH}/${WM_UTS_EXEC_NAME}"
+        chmod +x "${WM_UTS_INSTALL_PATH}/${WM_UTS_EXEC_NAME}"
 
         export PATH=$PATH:/home/${USER}/.local/bin
-        _set_build_number ${WM_UTS_SERVICE_HOME}/bin/wm-utils.sh
     fi
+    _set_build_number "${WM_UTS_SERVICE_HOME}/bin/wm-utils.sh"
 }
 
 function _set_build_number
@@ -167,9 +163,9 @@ function _set_build_number
     for _target in "${_targets[@]}";
     do
         echo "filling version ${BUILD_VERSION} on: ${_target}"
-        cp ${_target} ${_target}.tmp
-        sed -i "s/#FILLVERSION/$(date +%F) - ${BUILD_VERSION}/g" ${_target}
-        rm ${_target}.tmp
+        cp "${_target}" "${_target}.tmp"
+        sed -i "s/#FILLVERSION/$(date +%F) - ${BUILD_VERSION}/g" "${_target}"
+        rm "${_target}.tmp"
     done
 }
 
@@ -179,15 +175,15 @@ function _restore_build_number
     for _target in "${_targets[@]}"
     do
         echo "reseting: ${_target}"
-        cp ${_target}.tmp ${_target}
+        cp "${_target}.tmp" "${_target}"
     done
 }
 
 function _main
 {
     setup_defaults
-    rm -rf ${WM_UTS_SERVICE_HOME}
-    rm -rf ${WM_UTS_INSTALL_PATH}/${WM_UTS_EXEC_NAME}
+    rm -rf "${WM_UTS_SERVICE_HOME:?}"
+    rm -rf "${WM_UTS_INSTALL_PATH}/${WM_UTS_EXEC_NAME:?}"
 
     if [[ "${INSTALL_SYSTEM_PKG}" == "true" ]]
     then
